+++ WebApplication/src/App.vue
@@ -1,18 +1,43 @@
 <script>
+function applyBootStrapTheme() {
+    // hack to switch the theme: get the value from our custom css
+    let bootstrapTheme = getComputedStyle(document.documentElement).getPropertyValue('--bootstrap-theme');
+    console.log("-------------- Applying Bootstrap theme ...", bootstrapTheme);
+    if (bootstrapTheme) {
+        // and set it to the 'html' element
+        document.documentElement.setAttribute('data-bs-theme', bootstrapTheme);
+    } else {
+        console.warn("-------------- Applying Bootstrap theme not defined yet, retrying ...");
+        setTimeout(applyBootStrapTheme, 100);
+    }
+}
+
 export default {
     async created() {
         console.log("Creating App...");
 
         {// Load the CSS dynamically since it can be a custom css
-            console.log("Loading the customizable CSS variables ...");
+            console.log("Loading the defaults + custom CSS ...");
 
             let link = document.createElement('link');
             link.rel = 'stylesheet';
             link.type = 'text/css';
-            link.href = 'customizable/variables.css';
+            link.href = 'customizable/custom.css';
     
             document.getElementsByTagName('HEAD')[0].appendChild(link);
+
+            setTimeout(applyBootStrapTheme, 0);
         }
+        // {
+        //     console.log("Loading the custom CSS ...");
+
+        //     let link = document.createElement('link');
+        //     link.rel = 'stylesheet';
+        //     link.type = 'text/css';
+        //     link.href = 'customizable/custom.css';
+    
+        //     document.getElementsByTagName('HEAD')[0].appendChild(link);
+        // }
 
         await this.$store.dispatch('configuration/load');
         await this.$store.dispatch('studies/initialLoad');
@@ -40,7 +65,6 @@ export default {
     -webkit-font-smoothing: antialiased;
     -moz-osx-font-smoothing: grayscale;
     text-align: center;
-    color: #2c3e50;
 }
 
 .nav-side-layout {